all methods working for me on my machine; no buffer pool
have everything passing our old tests at the moment. only things i think
we ought to fix at the moment are making sure adjacent free blocks
connect and then implementing the buffer pool, which is gonna be a bear
because we didnt use the size of the message as a parameter. but we will
figure it out. :)