not passing all of the remove tests
failing two remove tests, but i think one of them is an error on their
part; they want a smaller file, but the file needs to be in increments
of the buffer size. so. thats not possible. then we are also outputting
something wrong because its not detecting a simple remove.